+++ textmerger/src-tauri/src/lib.rs
@@ -129,6 +129,7 @@ pub fn run() {
         .plugin(tauri_plugin_fs::init())
         .plugin(tauri_plugin_log::Builder::default().build())
         .plugin(tauri_plugin_dialog::init())
+        .plugin(tauri_plugin_shell::init())
         .invoke_handler(tauri::generate_handler![add_files, get_merged_content, exit_app])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");